added basic texture rendering, added integer timekeeping, added gamepad support

+++ include/util/clock.h
@@ -5,7 +5,7 @@
 typedef std::chrono::high_resolution_clock Clock;
 typedef float DeltaTime_t;
 typedef std::chrono::duration<DeltaTime_t> DeltaTime;
-typedef long ElapsedTime_t;
+typedef long long ElapsedTime_t;
 typedef std::chrono::duration<ElapsedTime_t> ElapsedTime;
 
 #endif // CLOCK_H
